Add pom information to all runtime artifacts, fix naming
Ensure the required pom information is added to every runtime artifact. Also do some magic to ensure all of the runtime names are the same. Source jars are added to all platform artifacts now as of 1.3 RC2, we still need to add an empty javadoc jar to each platform though.