Move downloading intellij outside configuration.
Downloading and unzipping of intellij were previously declared inside of
the body of a task instead of during the task action which meant that we
tried to download and unzip intellij whenever any gradle task was
executed.

This changes the download and unzip tasks to be separate tasks and
allows up to date checking to work properly. This reduces the
configuration time of the gradle build by about 15 seconds when the
files are already downloaded and unzipped.