Use relative path sensitivity for SourceTask.
The absolute path is not relevant in this case when caching since we
only care about the relative path of the files inside the project.

According to the docs, this is the proper way of configuring the
PathSensitivity for the SourceTask.

https://docs.gradle.org/current/dsl/org.gradle.api.tasks.SourceTask.html

```
The PathSensitivity for the sources is configured to be PathSensitivity.ABSOLUTE. If your sources are less strict, please change it accordingly by overriding this method in your subclass.
```