Fix inconsistency in directory name sanitizing (#1650)
Before, if a project was named `taco-db`, the implementation package would correctly sanitize it to `tacodb` for the package name. The `FileIndex` used to generate later references to the generated implementation didn't do this same directory name sanitizing though, so it would try to generate a a package name with ``taco-db`` in it, which wasn't right and would fail to compile.

This tries to unify their behavior with a shared function.