Use nullable argument if compared using IS NULL at any point.
Currently the nullability of a parameter is determined by the
ordering of where it comes in the statement. This change aims
to make it independent of ordering, such that these two queries
have the same argument types:

SELECT * FROM data WHERE token = :token OR :token IS NULL;
SELECT * FROM data WHERE :token IS NULL OR token = :token;

If we assume data.token is TEXT, both queries should have a
parameter "token" of type `kotlin.String?`.