Do not depend on kotlin-native-utils in Gradle plugin
kotlin-native-utils.jar is an internal artifact published due to
technical reasons only. But now there is no need for it anymore
and its publication will be stopped.

This patch removes a dependency on this artifact. It also removes
usages of the KonanTarget class because this class is an internal
detail of the Kotlin compiler which can be changed at any moment.