Stop invoking `forDatabaseFiles` callback on directories
It's possible that a consuming codebase could have a directory
with the suffix `.db`. Therefore, SQLDelight would assume that
a migration was being performed, and execute the task
`verifyMainDatabaseMigration`. Since the "file" that was collected
was actually a directory, and not a database file, the task would
fail by trying to parse a number from the prefix contents of
the directory name.

This commit simplifies database file collection using Kotlin's
FileTreeWalk extension function `File.walk`. We ensure that
callbacks are invoked only for a File if `File.isFile` is true.

References #2381.