Update migrations.md
On iOS we were experiencing a crash because we had added
BEGIN/END TRANSACTION to our migrations. The error message indicated
that a transaction had been nested – we checked the documentation and
didn't find any mention of transactions running in a transaction, so
it seemed like a good idea to mention it here.