Provide SqlDriver as a parameter to AfterVersion
In order to run an update, a SqlDriver is required in order to actually
run update queries within an AfterVersion. Due to the fact that the
callback is provided as a parameter to the AndroidSqliteDriver
constructor, this makes it difficult to get the database to do updates
without wrapping the existing AndroidSqliteDriver.Callback or a
SupportSQLiteOpenHelper.Callback instead. This patch adds the SqlDriver
as a parameter to the AfterVersion lambda to make this easier.

Fixes #2534 and fixes #2614.