Build tweaks (#2437)
* Apply plugins with plugins DSL where possible

* Use GrammarKit Gradle plugin 2021.1.3

This includes fixes for deprecated Gradle features

* Avoid resolving compileOnly configuration

This is deprecated in Gradle

* Use in-memory database for Android Paging 3 extension tests

* Disable buildSearchableOptions task

This task builds an index of UI components (a.k.a. searchable options) for
the plugin. This is unnecessary as there is no UI for the SqlDelight
IntelliJ plugin. Disabling the task saves time during each build.

* Remove redundant config

Dependencies script is already applied by buildscript.gradle

* Stop copying Gradle Wrapper for tests

TestKit sets up the Gradle environment so this is unnecessary

* Update Robolectric to 4.7.2

Supports Java 17

* Skip linking for Linux test build for coroutines extensions

This fails on Windows

* Remove unused configuration

This was previously used to inject plugins for TestKit-based tests which
aren't used any longer.