Add common types to JdbcDriver
This enables dialects that use the JDBC driver to access JDBC APIs directly instead of needing to convert to one of the SQLite types.

These bindings should be fairly universal amongst JDBC drivers. In cases where they might not be desirable, such as SQLite, they don't need to be used. Each dialect has to opt-in to calling the methods. Date/time types are not included as I believe different drivers have different opinions on using the types from java.time vs java.sql, and the date/time APIs have more variance. This is something that can be improved as we see driver implementation arrive.

#2821 #2128