Customize codegen based on dialect (#2918)
* Create interfaces needed in compiler to implement codegen

These new interfaces allow the codegen to be customized on a per-dialect basis. This enables different dialects to further customize the implementation. For example, H2 can use JDBC functionality directly, and Postgres can use custom types #2821.

Further work is needed to modify the Gradle plugin to make these dialects selectable and even customizable.

* Customize codegen based on dialect

With this change, the codegen produces code that is customized for each dialect. This allows for dialects to use dialect-specific types and patterns.

#2821

* fixup! Customize codegen based on dialect

* fixup! Customize codegen based on dialect

* fixup! Customize codegen based on dialect

* fixup! Create interfaces needed in compiler to implement codegen

* fixup! Customize codegen based on dialect

* fixup! Create interfaces needed in compiler to implement codegen

* fixup! Customize codegen based on dialect

* fixup! Customize codegen based on dialect

* fixup! Create interfaces needed in compiler to implement codegen

* fixup! Customize codegen based on dialect

* fixup! Create interfaces needed in compiler to implement codegen

* fixup! Customize codegen based on dialect

* fixup! Create interfaces needed in compiler to implement codegen

* fixup! Customize codegen based on dialect