Fix codegen for RETURNING clause (#3872)
* RETURNING multiple values on INSERT, UDPATE, DELETE generates wrong code

The code generation is correct for scenarios where a single column is returned, as well as when all columns are returned. However, if multiple, but not all, columns are returned, then the generated code still uses the table type, rather than generating a new interface for the projection.

* Fix: RETURNING multiple values on INSERT, UPDATE, DELETE generates wrong code

The code generation was correct for scenarios where a single column is returned, as well as when all columns are returned. However, if multiple, but not all, columns are returned, then the generated code still used the table type, rather than generating a new interface for the projection.

With this change, the pureTable computation more closely resembles that of SelectQueryable, which contains more complex logic for purity identification.