Add Character length functions for PostgreSql Dialect
Returns number of characters in the string or will return NULL if expression is NULL.
In PostgreSql the following are synonymous and are Ansi Sql supported.
length ( In MySql the length function is in bytes )
char_length
character_length

char_length is ANSI SQL Standard and the explicitly named character_length,
as multibyte character counts as a single character not as bytes.

Examples

char_length ( text ) -> integer
char_length('jose') -> 4
char_length('海豚') -> 2
char_length(NULL) -> NULL