Add Hsql char length function
Hsql supports LENGTH, CHAR_LENGTH, CHARACTER_LENGTH and are synonymous.

All functions return a BIGINT value that measures the length of the string in characters.

There appeared to be an issue in running HsqlTest where the schema is created each test.
Appending “;shutdown=true” to the connection string closes the in memory database.