Add Language annotations for more pleasant SQL editing
Using the execute and executeQuery methods isn't how SQLDelight is meant to be used, but they are an important escape hatch or when the compiler/dialect hasn't yet implemented a needed feature. Marking the parameters as SQL makes for a better editing experience in Intellij.

The required annotations are already on the classpath for these drivers.