RegExMatchOperatorExpressionMixin
Annotation errors for using regex operator with a non-text data type
Change message to better understanding

The problem here is that Functions and Column can be used with Regex operators however, there appears to be no way
to access the typeresolver to get the function return type.  We can only validate Column Types and functions we have to let through.

TypeResolvers are in core compiler package no available here