Make JVM samples consistent with other platforms
The JVM code samples do not use the `PRAGMA user_version`-based
migration logic, which is a bit misleading because samples for other
platforms do.

The inconsistency is particularly visible in the Multiplatform case,
where the android and native samples will set `user_version=1`, but the
JVM sample will produce `user_version=0`.