Generated query files return row counts for simple mutators (#4578)
* Generated query files return row counts for simple mutators

Simple mutators like INSERT, UPDATE, and DELETE, without RETURNING clauses, typically return the number of rows changed. In fact, the SqlDriver#execute() function already does this. However, the value of execute() is not exposed to the generated query files.

This commit makes it so that simple INSERT, UPDATE, and DELETE statements return the QueryResult<Long> that the SqlDriver already returns.

* Spotless

* Fix tests

* Update integration tests

---------

Co-authored-by: Alec Kazakova <1675456+AlecKazakova@users.noreply.github.com>
Co-authored-by: Alec Kazakova <aleckazakova@squareup.com>