Fix Sqlite FTS5 data type requirement (#5896)
* Update TreeUtil.kt

Remove unused `SqlCreateVirtualTableStmt` as this is never a parent  of SqlColumnName

Add `SqlModuleColumnDef` as this is only matched when there is no column data type - assumes TEXT nullable

* Update InterfaceGeneration.kt

Compiler test for `VIRTUAL TABLE Email USING fts` when referencing columns in sql body.
These are compiled to nullable String types.

This is the same behaviour if data types are specified.

* Update CHANGELOG.md

Co-authored-by: Jake Wharton <github@jakewharton.com>