Add common table expression WITH ... AS [NOT] MATERIALIZED (#5961)
* Add WITH ... AS [NOT MATERIALIZED ] Postgresql

MATERIALIZED is used to inline the CTE into the top level query.

The query planner may or may not use this optimization.

* Add WITH ... [ NOT ] MATERIALIZED

Sqlite v3.35 copied this from PostgreSql

MATERIALIZED is used to inline the CTE into the top level query.

The query planner may or may not use this optimization.

* Update CHANGELOG.md