Improve error message for CAST compiler error (#5979)
* Update TypeResolver.kt

AnnotationException is an IllegalStateException and can work with existing code.
It takes a PsiElement to locate the bind argument in the source file used in error message formatIng.
This is performed by SqlDelightEnvironment

* Update SqlDelightEnvironment.kt

Catch any AnnotationException from resolvers when compiler is writing interfaces.
Currently only parsing of source produced annotation errors

* Update CHANGELOG.md

* Add gradle task test

Needed for error message check