Allow tags and snapshots to publish concurrently (#6002)
Previously, when doing a release, the commit gets pushed up before the tag. So this concurrency grouping prevented the tag from publishing while we wait for the commit to publish the next snapshot. This should separate the branch and the tag and allow them to run at the same time.